# doc/openapi.yaml
openapi: 3.0.2
info:
  title: Notes API
  description: Notes API description.
  version: "v1.0.0"
servers:
  - url: https://notes.sp/api
    description: Development server
components:
  responses:
...
    Forbidden:
      description: Access denied
...
  requestBodies:
...
    SignIn:
      description: SignIn request
      content:
        application/json:
          schema:
            type: object
            properties:
              identity:
                type: string
              credential:
                type: string
                format: password
...
  securitySchemes:
    Authorization:
      type: apiKey
      in: header
      name: Authorization
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
...
paths:
...
  /sign-in:
    post:
      summary: Sing In.
      operationId: signIn
      security:
        - ApiKey: []
      requestBody:
        $ref: '#/components/requestBodies/SignIn'
      responses:
        200:
          description: Responds with auth token
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
        403:
          $ref: '#/components/responses/Forbidden'
...
  /notes:
    post:
      summary: Create a service
      operationId: addService
      security:
        - ApiKey: []
        - Authorization: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Service'
      responses:
        201:
          description: Responds with the created Service object
          headers:
            Location:
              description: URI of the created Service
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Service'
        403:
          $ref: '#/components/responses/Forbidden'
...